pyproject.toml
[tool.poetry]
name = "repl_python3_Python-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.17.1"
youtube-dl = "^2021.5.16"
wavelink = "^0.9.9"
pymongo = "^3.11.4"
tinydb = "^4.4.0"
PyNaCl = "^1.4.0"
discord-ext-slash = "^0.7.0"
discord-py-slash-command = "^2.3.2"
Flask = "^2.0.1"
lab-notebook = "^3.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
